Reorganise the library for date picker and fix flaky tests for loading pages
