+++ tests/support/page-object-model/login-page.ts
@@ -24,7 +24,7 @@ export class LoginPage {
 
      /* Open the page login */
      async goto(){
-        await this.page.goto('https://demoqa.com/login');
+        await this.page.goto('https://demoqa.com/login',{waitUntil: 'load'});
 
      }
