+++ tests/testcases/resizable.spec.ts
@@ -3,7 +3,7 @@ import { Resizable } from '../support/page-object-model/resizable';
 
 /* @Author: Thu Nguyen */
 test.beforeEach(async ({ page }) => {
-  await page.goto('https://demoqa.com/resizable');
+  await page.goto('https://demoqa.com/resizable',{waitUntil: 'load'});
 });
 
 test.describe('Resize the boxes',()=> {
