+++ tests/testcases/slider.spec.ts
@@ -3,7 +3,7 @@ import { Slider } from '../support/page-object-model/slider';
 
 /* @Author: Thu Nguyen */
 test.beforeEach(async ({ page }) => {
-  await page.goto('https://demoqa.com/slider');
+  await page.goto('https://demoqa.com/slider',{waitUntil: 'load'});
 });
 
 test.describe('Drag the slider',()=> {
